Tidy up project
