+++ main.py
@@ -1,5 +1,6 @@
-from sentimental import web_scraper as ws
+from sentimental import web_scraper
 
-scraper = ws.GoogleScraper('sp500', 'MarketWatch')
-data = scraper.scrape_n(10, wait=0)
-print(data)
+if __name__ == "__main__":
+    scraper = web_scraper.GoogleScraper('sp500', 'MarketWatch')
+    data = scraper.scrape_n(10, wait=0)
+    print(data)
